Bugs fixed, refactor code, will add toast, remove useless ifs add new design pattern
