+++ src/pages/Login/index.js
@@ -6,39 +6,25 @@ import { AppButton, ContentView, Form, Page } from '../../styles/default';
 import Footer from '../../components/Footer';
 import api from '../../services/api';
 import { login } from '../../services/auth';
-
+import history from '../../services/history/history'
 function Login() {
 
-  const [buttonText, setButtonText] = useState("Login");
 
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [errors, setErrors] = useState({});
+
 
   const handleLoginSubmit = async (e) => {
     e.preventDefault();
-    setButtonText("Enviando Dados ...");
 
     try {
       const response = await api.post("/auth/authenticate", {email, password});
 
       login(response.data.token, response.data.user);
-      setButtonText("Logado com Sucesso");
+      history.push('/')
     } catch (error) {
-      setButtonText("Tente Novamente");
+      //Toastify over here
 
-      if(error.response.data) {
-        //Dados retornados do backend
-        if(error.response.data.errors) {
-          setErrors(error.response.data.errors);
-        }
-        if(error.response.data.message) {
-          setErrors({message: error.response.data.message});
-        }
-      } else {
-        //Não houve dados retornados do backend
-        alert("Erro Inesperado!");
-      }
     }
   };
 
@@ -47,35 +33,28 @@ function Login() {
     <Header/>
     <Form width={"45%"} center>
         <ContentView>
-          <label>Faça Login na Nossa Plataforma!</label>
+          <label>Hola adm! inicia sesión en nuestra plataforma</label>
 
-          <label style={{color: 'red'}}>{errors.message}</label>
-
+    <br></br>
           <input
-            placeholder="Insira seu E-mail"
+            placeholder="E-mail"
             type="email"
             onChange={(e) => {
               setEmail(e.target.value);
             }}
             value={email}
           />
-          <span style={{color: 'red'}}>{errors.email}</span>
 
           <input
-            placeholder="Insira sua Senha "
+            placeholder="Contraseña "
             type="password"
              onChange={(e) => {
               setPassword(e.target.value);
             }}
             value={password}
           />
-          <span style={{color: 'red'}}>{errors.password}</span>
-
-          <Link to="/cadastro">
-            Ainda não Tem Conta? Cadastre-se Clicando Aqui!!
-          </Link>
-          <br></br>
-          <AppButton onClick={handleLoginSubmit}>{buttonText}</AppButton>
+    <br></br>
+          <AppButton onClick={handleLoginSubmit}>Iniciar Sesión</AppButton>
         </ContentView>
     </Form>
     <Footer/>
